+++ xraptor/__init__.py
@@ -3,6 +3,15 @@ from .broadcaster import Broadcast
 from .domain.methods import MethodType
 from .domain.request import Request
 from .domain.response import Response
+from .observability import Metrics
 from .server import XRaptor
 
-__all__ = ["Broadcast", "MethodType", "Request", "Response", "XRaptor", "antennas"]
+__all__ = [
+    "Broadcast",
+    "MethodType",
+    "Metrics",
+    "Request",
+    "Response",
+    "XRaptor",
+    "antennas",
+]